ver0.3 plus writer/see menu/change profileImage and name

+++ controllers/boardsControllers.js
@@ -1,9 +1,10 @@
 import Board from "../models/Boards.js";
 import routes from "../routes.js";
+import User from "../models/User.js";
 
 export const boards = async (req, res) => {
   try {
-    const list = await Board.find({}).sort({createdAt : -1});
+    const list = await Board.find({}).sort({ createdAt: -1 });
     res.render("boards", { pageName: "게시판", list });
   } catch (error) {
     console.log(error);
@@ -14,42 +15,64 @@ export const boards = async (req, res) => {
 export const getBoardsUpload = (req, res) => {
   res.render("boardsUpload", { pageName: "글쓰기", Board });
 };
-  
+
 export const postBoardsUpload = async (req, res) => {
   const {
     body: { title, description },
+    user,
   } = req;
+
+  // let array = [];
 
   const newBoards = await Board.create({
     title,
     description,
+    creator: user._id,
   });
-  console.log(newBoards);
-  res.redirect(routes.boardsDetail(newBoards.id));
+
+  // user.boards.forEach((boards) => array.push(boards));
+  // array.push(String(newBoards._id));
+
+  // await User.findOneAndUpdate({ _id: user._id }, { board: array });
+
+  res.redirect(routes.boardsDetail(newBoards._id));
 };
 
 export const boardDetail = async (req, res) => {
   const {
-    body: { title },
     params: { id },
   } = req;
+
   try {
-    const boards = await Board.findById(id);
+    const boards = await Board.findById(id).populate("creator");
+
     res.render("boardsDetail", {
       pageName: "게시판",
       id,
       title: boards.title,
       description: boards.description ? boards.description : [],
+      views: boards.views,
+      createdAt: boards.createdAt,
+      creator: boards.creator,
     });
   } catch (error) {}
 };
 
-export const getBoardsEdit = (req, res) => {
+export const getBoardsEdit = async (req, res) => {
   const {
     params: { id },
   } = req;
 
-  res.render("boardsEdit", { pageName: "수정하기", id });
+  try {
+    const board = await Board.findById(id);
+    if (String(board.creator) !== req.user._id) {
+      throw Error();
+    } else {
+      res.render("boardsEdit", { pageName: "수정하기", id, board });
+    }
+  } catch (error) {
+    console.log(error);
+  }
 };
 
 export const postBoardsEdit = async (req, res) => {
@@ -82,7 +105,12 @@ export const boardsDelete = async (req, res) => {
   const {
     params: { id },
   } = req;
+
   try {
+    const board = await Board.findById(id);
+    if (String(board.creator) !== req.user._id) {
+      throw Error();
+    }
     await Board.findOneAndRemove({ _id: id });
   } catch (error) {
     console.log(error);
